+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/graph/BindingGraphTest.kt
@@ -5,13 +5,12 @@ package dev.zacsweers.metro.compiler.graph
 import androidx.collection.ScatterMap
 import com.google.common.truth.Truth.assertThat
 import dev.zacsweers.metro.compiler.tracing.TraceScope
-import dev.zacsweers.metro.compiler.tracing.emptyTraceScope
 import kotlin.test.assertFails
 import kotlin.test.assertFailsWith
 import kotlin.test.assertTrue
 import org.junit.Test
 
-class BindingGraphTest : TraceScope by emptyTraceScope("test") {
+class BindingGraphTest : TraceScope by TraceScope.noop() {
 
   @Test
   fun put() {